use-cases/agent-factory-with-subagents/agents/idea_brainstorm_agent/requirements.txt
pydantic-ai-slim[anthropic,openai]==0.7.4
pydantic>=2.12
pydantic-settings>=2.11
python-dotenv>=1.2
rich>=14.2
pytest>=8.4
